Implemented query caching for searches
This commit implements a possibly multi-user cache for long running
queries (currently the threshold is 0.4 seconds)
We use the max (message id, topic id) as a marker in the cache keys,
because these lookups are cheap. This means the cache remains valid for
multiple users as long as no new message arrives matching the search
criteria, and at time we automatically generate a new cache entry, while
existing views can still use the old cache record.